+++ packages/gatsby-recipes/src/create-types.test.js
@@ -4,6 +4,42 @@ test(`create-types`, () => {
   const result = createTypes()
   expect(result.mutationTypes).toMatchInlineSnapshot(`
     Object {
+      "createContentfulEntry": Object {
+        "args": Object {
+          "contentfulEntry": Object {
+            "type": "ContentfulEntryInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEntry",
+      },
+      "createContentfulEnvironment": Object {
+        "args": Object {
+          "contentfulEnvironment": Object {
+            "type": "ContentfulEnvironmentInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEnvironment",
+      },
+      "createContentfulSpace": Object {
+        "args": Object {
+          "contentfulSpace": Object {
+            "type": "ContentfulSpaceInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulSpace",
+      },
+      "createContentfulType": Object {
+        "args": Object {
+          "contentfulType": Object {
+            "type": "ContentfulTypeInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulType",
+      },
       "createDirectory": Object {
         "args": Object {
           "directory": Object {
@@ -22,6 +58,15 @@ test(`create-types`, () => {
         "resolve": [Function],
         "type": "File",
       },
+      "createGatsbyPage": Object {
+        "args": Object {
+          "gatsbyPage": Object {
+            "type": "GatsbyPageInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbyPage",
+      },
       "createGatsbyPlugin": Object {
         "args": Object {
           "gatsbyPlugin": Object {
@@ -39,6 +84,15 @@ test(`create-types`, () => {
         },
         "resolve": [Function],
         "type": "GatsbyShadowFile",
+      },
+      "createGatsbySiteMetadata": Object {
+        "args": Object {
+          "gatsbySiteMetadata": Object {
+            "type": "GatsbySiteMetadataInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbySiteMetadata",
       },
       "createGitIgnore": Object {
         "args": Object {
@@ -76,6 +130,42 @@ test(`create-types`, () => {
         "resolve": [Function],
         "type": "NPMScript",
       },
+      "destroyContentfulEntry": Object {
+        "args": Object {
+          "contentfulEntry": Object {
+            "type": "ContentfulEntryInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEntry",
+      },
+      "destroyContentfulEnvironment": Object {
+        "args": Object {
+          "contentfulEnvironment": Object {
+            "type": "ContentfulEnvironmentInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEnvironment",
+      },
+      "destroyContentfulSpace": Object {
+        "args": Object {
+          "contentfulSpace": Object {
+            "type": "ContentfulSpaceInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulSpace",
+      },
+      "destroyContentfulType": Object {
+        "args": Object {
+          "contentfulType": Object {
+            "type": "ContentfulTypeInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulType",
+      },
       "destroyDirectory": Object {
         "args": Object {
           "directory": Object {
@@ -94,6 +184,15 @@ test(`create-types`, () => {
         "resolve": [Function],
         "type": "File",
       },
+      "destroyGatsbyPage": Object {
+        "args": Object {
+          "gatsbyPage": Object {
+            "type": "GatsbyPageInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbyPage",
+      },
       "destroyGatsbyPlugin": Object {
         "args": Object {
           "gatsbyPlugin": Object {
@@ -111,6 +210,15 @@ test(`create-types`, () => {
         },
         "resolve": [Function],
         "type": "GatsbyShadowFile",
+      },
+      "destroyGatsbySiteMetadata": Object {
+        "args": Object {
+          "gatsbySiteMetadata": Object {
+            "type": "GatsbySiteMetadataInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbySiteMetadata",
       },
       "destroyGitIgnore": Object {
         "args": Object {
@@ -148,6 +256,42 @@ test(`create-types`, () => {
         "resolve": [Function],
         "type": "NPMScript",
       },
+      "updateContentfulEntry": Object {
+        "args": Object {
+          "contentfulEntry": Object {
+            "type": "ContentfulEntryInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEntry",
+      },
+      "updateContentfulEnvironment": Object {
+        "args": Object {
+          "contentfulEnvironment": Object {
+            "type": "ContentfulEnvironmentInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEnvironment",
+      },
+      "updateContentfulSpace": Object {
+        "args": Object {
+          "contentfulSpace": Object {
+            "type": "ContentfulSpaceInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulSpace",
+      },
+      "updateContentfulType": Object {
+        "args": Object {
+          "contentfulType": Object {
+            "type": "ContentfulTypeInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulType",
+      },
       "updateDirectory": Object {
         "args": Object {
           "directory": Object {
@@ -166,6 +310,15 @@ test(`create-types`, () => {
         "resolve": [Function],
         "type": "File",
       },
+      "updateGatsbyPage": Object {
+        "args": Object {
+          "gatsbyPage": Object {
+            "type": "GatsbyPageInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbyPage",
+      },
       "updateGatsbyPlugin": Object {
         "args": Object {
           "gatsbyPlugin": Object {
@@ -183,6 +336,15 @@ test(`create-types`, () => {
         },
         "resolve": [Function],
         "type": "GatsbyShadowFile",
+      },
+      "updateGatsbySiteMetadata": Object {
+        "args": Object {
+          "gatsbySiteMetadata": Object {
+            "type": "GatsbySiteMetadataInput",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbySiteMetadata",
       },
       "updateGitIgnore": Object {
         "args": Object {
@@ -224,9 +386,33 @@ test(`create-types`, () => {
   `)
   expect(result.queryTypes).toMatchInlineSnapshot(`
     Object {
+      "allContentfulEntry": Object {
+        "resolve": [Function],
+        "type": "ContentfulEntryConnection",
+      },
+      "allContentfulEnvironment": Object {
+        "resolve": [Function],
+        "type": "ContentfulEnvironmentConnection",
+      },
+      "allContentfulSpace": Object {
+        "resolve": [Function],
+        "type": "ContentfulSpaceConnection",
+      },
+      "allContentfulType": Object {
+        "resolve": [Function],
+        "type": "ContentfulTypeConnection",
+      },
+      "allGatsbyPage": Object {
+        "resolve": [Function],
+        "type": "GatsbyPageConnection",
+      },
       "allGatsbyPlugin": Object {
         "resolve": [Function],
         "type": "GatsbyPluginConnection",
+      },
+      "allGatsbySiteMetadata": Object {
+        "resolve": [Function],
+        "type": "GatsbySiteMetadataConnection",
       },
       "allGitIgnore": Object {
         "resolve": [Function],
@@ -239,6 +425,42 @@ test(`create-types`, () => {
       "allNPMScript": Object {
         "resolve": [Function],
         "type": "NPMScriptConnection",
+      },
+      "contentfulEntry": Object {
+        "args": Object {
+          "id": Object {
+            "type": "String",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEntry",
+      },
+      "contentfulEnvironment": Object {
+        "args": Object {
+          "id": Object {
+            "type": "String",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulEnvironment",
+      },
+      "contentfulSpace": Object {
+        "args": Object {
+          "id": Object {
+            "type": "String",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulSpace",
+      },
+      "contentfulType": Object {
+        "args": Object {
+          "id": Object {
+            "type": "String",
+          },
+        },
+        "resolve": [Function],
+        "type": "ContentfulType",
       },
       "directory": Object {
         "args": Object {
@@ -258,6 +480,15 @@ test(`create-types`, () => {
         "resolve": [Function],
         "type": "File",
       },
+      "gatsbyPage": Object {
+        "args": Object {
+          "id": Object {
+            "type": "String",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbyPage",
+      },
       "gatsbyPlugin": Object {
         "args": Object {
           "id": Object {
@@ -275,6 +506,15 @@ test(`create-types`, () => {
         },
         "resolve": [Function],
         "type": "GatsbyShadowFile",
+      },
+      "gatsbySiteMetadata": Object {
+        "args": Object {
+          "id": Object {
+            "type": "String",
+          },
+        },
+        "resolve": [Function],
+        "type": "GatsbySiteMetadata",
       },
       "gitIgnore": Object {
         "args": Object {