+++ packages/gatsby/src/services/create-pages-statefully.ts
@@ -5,6 +5,7 @@ import { IDataLayerContext } from "../state-machines/data-layer/types"
 export async function createPagesStatefully({
   parentSpan,
   gatsbyNodeGraphQLFunction,
+  deferNodeMutation,
 }: Partial<IDataLayerContext>): Promise<void> {
   // A variant on createPages for plugins that want to
   // have full control over adding/removing pages. The normal
@@ -21,7 +22,7 @@ export async function createPagesStatefully({
       traceId: `initial-createPagesStatefully`,
       waitForCascadingActions: true,
       parentSpan: activity.span,
-      // deferNodeMutation: true, //later
+      deferNodeMutation,
     },
     {
       activity,