+++ packages/gatsby/src/services/start-webpack-server.ts
@@ -3,6 +3,7 @@ import report from "gatsby-cli/lib/reporter"
 import formatWebpackMessages from "react-dev-utils/formatWebpackMessages"
 import chalk from "chalk"
 import { Compiler } from "webpack"
+import { isEqual } from "lodash"
 
 import {
   reportWebpackWarnings,
@@ -12,7 +13,7 @@ import {
 import { printDeprecationWarnings } from "../utils/print-deprecation-warnings"
 import { printInstructions } from "../utils/print-instructions"
 import { prepareUrls } from "../utils/prepare-urls"
-import { startServer } from "../utils/start-server"
+import { startServer, IWebpackWatchingPauseResume } from "../utils/start-server"
 import { WebsocketManager } from "../utils/websocket-manager"
 import { IBuildContext } from "./"
 import {
@@ -20,36 +21,50 @@ import {
   markWebpackStatusAsDone,
 } from "../utils/webpack-status"
 import { enqueueFlush } from "../utils/page-data"
+import mapTemplatesToStaticQueryHashes from "../utils/map-templates-to-static-query-hashes"
+import { emitter } from "../redux"
 
 export async function startWebpackServer({
   program,
   app,
   workerPool,
+  store,
 }: Partial<IBuildContext>): Promise<{
   compiler: Compiler
   websocketManager: WebsocketManager
+  webpackWatching: IWebpackWatchingPauseResume
 }> {
-  if (!program || !app) {
-    throw new Error(`Missing required params`)
+  if (!program || !app || !store) {
+    report.panic(`Missing required params`)
   }
-  let { compiler, webpackActivity, websocketManager } = await startServer(
-    program,
-    app,
-    workerPool
-  )
+  let {
+    compiler,
+    webpackActivity,
+    websocketManager,
+    webpackWatching,
+  } = await startServer(program, app, workerPool)
+  webpackWatching.suspend()
 
   compiler.hooks.invalid.tap(`log compiling`, function () {
-    markWebpackStatusAsPending()
+    if (!webpackActivity) {
+      // mark webpack as pending if we are not in the middle of compilation already
+      // when input is invalidated during compilation, webpack will automatically
+      // run another compilation round before triggering `done` event
+      report.pendingActivity({ id: `webpack-develop` })
+      markWebpackStatusAsPending()
+    }
   })
 
   compiler.hooks.watchRun.tapAsync(`log compiling`, function (_, done) {
-    if (webpackActivity) {
-      webpackActivity.end()
+    if (!webpackActivity) {
+      // there can be multiple `watchRun` events before receiving single `done` event
+      // webpack will not emit assets or `done` event until all pending invalidated
+      // inputs were compiled
+      webpackActivity = report.activityTimer(`Re-building development bundle`, {
+        id: `webpack-develop`,
+      })
+      webpackActivity.start()
     }
-    webpackActivity = report.activityTimer(`Re-building development bundle`, {
-      id: `webpack-develop`,
-    })
-    webpackActivity.start()
 
     done()
   })
@@ -109,10 +124,46 @@ export async function startWebpackServer({
         webpackActivity.end()
         webpackActivity = null
       }
-      enqueueFlush()
+
+      if (isSuccessful) {
+        const state = store.getState()
+        const mapOfTemplatesToStaticQueryHashes = mapTemplatesToStaticQueryHashes(
+          state,
+          stats.compilation
+        )
+
+        mapOfTemplatesToStaticQueryHashes.forEach(
+          (staticQueryHashes, componentPath) => {
+            if (
+              !isEqual(
+                state.staticQueriesByTemplate.get(componentPath),
+                staticQueryHashes
+              )
+            ) {
+              store.dispatch({
+                type: `ADD_PENDING_TEMPLATE_DATA_WRITE`,
+                payload: {
+                  componentPath,
+                },
+              })
+              store.dispatch({
+                type: `SET_STATIC_QUERIES_BY_TEMPLATE`,
+                payload: {
+                  componentPath,
+                  staticQueryHashes,
+                },
+              })
+            }
+          }
+        )
+
+        enqueueFlush()
+      }
+
       markWebpackStatusAsDone()
       done()
-      resolve({ compiler, websocketManager })
+      emitter.emit(`COMPILATION_DONE`, stats)
+      resolve({ compiler, websocketManager, webpackWatching })
     })
   })
 }