+++ www/src/utils/node/docs.js
@@ -1,31 +1,6 @@
-const minimatch = require(`minimatch`)
-
 const { getPrevAndNext } = require(`../get-prev-and-next.js`)
 const { getMdxContentSlug } = require(`../get-mdx-content-slug`)
 const { getTemplate } = require(`../get-template`)
-const findApiCalls = require(`../find-api-calls`)
-
-const ignorePatterns = [
-  `**/commonjs/**`,
-  `**/node_modules/**`,
-  `**/__tests__/**`,
-  `**/dist/**`,
-  `**/__mocks__/**`,
-  `babel.config.js`,
-  `graphql.js`,
-  `**/flow-typed/**`,
-]
-
-function isCodeFile(node) {
-  return (
-    node.internal.type === `File` &&
-    node.sourceInstanceName === `gatsby-core` &&
-    [`js`].includes(node.extension) &&
-    !ignorePatterns.some(ignorePattern =>
-      minimatch(node.relativePath, ignorePattern)
-    )
-  )
-}
 
 function mdxResolverPassthrough(fieldName) {
   return async (source, args, context, info) => {
@@ -71,24 +46,6 @@ exports.createSchemaCustomization = ({ actions: { createTypes } }) => {
       tableOfContents: JSON
       excerpt: String!
     }
-
-    type GatsbyAPICall implements Node @derivedTypes @dontInfer {
-      name: String
-      file: String
-      group: String
-      codeLocation: GatsbyAPICallCodeLocation
-    }
-
-    type GatsbyAPICallCodeLocation @dontInfer {
-      filename: Boolean
-      end: GatsbyAPICallEndpoint
-      start: GatsbyAPICallEndpoint
-    }
-
-    type GatsbyAPICallEndpoint @dontInfer {
-      column: Int
-      line: Int
-    }
   `)
 }
 
@@ -115,33 +72,10 @@ exports.onCreateNode = async ({
   node,
   actions,
   getNode,
-  loadNodeContent,
   createNodeId,
   createContentDigest,
 }) => {
   const { createNode, createParentChildLink } = actions
-
-  if (isCodeFile(node)) {
-    const calls = await findApiCalls({ node, loadNodeContent })
-    if (calls.length > 0) {
-      calls.forEach(call => {
-        const apiCallNode = {
-          id: createNodeId(`findApiCalls-${JSON.stringify(call)}`),
-          parent: node.id,
-          children: [],
-          ...call,
-          internal: {
-            type: `GatsbyAPICall`,
-          },
-        }
-        apiCallNode.internal.contentDigest = createContentDigest(apiCallNode)
-
-        createNode(apiCallNode)
-        createParentChildLink({ parent: node, child: apiCallNode })
-      })
-    }
-    return
-  }
 
   const slug = getMdxContentSlug(node, getNode(node.parent))
   if (!slug) return