+++ packages/gatsby-plugin-typescript/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.17...gatsby-plugin-typescript@2.4.18) (2020-08-10)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.4.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.16...gatsby-plugin-typescript@2.4.17) (2020-08-06)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.4.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.15...gatsby-plugin-typescript@2.4.16) (2020-07-20)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.4.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.14...gatsby-plugin-typescript@2.4.15) (2020-07-15)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
+## [2.4.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.13...gatsby-plugin-typescript@2.4.14) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-plugin-typescript
+
 ## [2.4.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typescript@2.4.12...gatsby-plugin-typescript@2.4.13) (2020-07-03)
 
 **Note:** Version bump only for package gatsby-plugin-typescript